Split up search_parameter_parser.rb
search_parameter_parser.rb had grown too large (629LOC). Breaking it up
in files per class, in a separate directly will help people better
understand the structure of the code.
