+++ app.rb
@@ -14,7 +14,8 @@ require "elasticsearch/index"
 require "elasticsearch/search_server"
 require "redis"
 require "matcher_set"
-require "search_parameter_parser"
+require "parameter_parser/search_parameter_parser"
+require "parameter_parser/facet_parameter_parser"
 require "registries"
 require "schema/combined_index_schema"
