+++ test/unit/parameter_parser/search_parameter_parser_test.rb
@@ -1,5 +1,6 @@
 require "test_helper"
-require "search_parameter_parser"
+require "parameter_parser/search_parameter_parser"
+require "parameter_parser/facet_parameter_parser"
 
 class SearchParameterParserTest < ShouldaUnitTestCase
